# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "one piece 3mk",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Localhost API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve a list of products with optional search and filtering",
                "operationId": "a2f6016e5bef16a083b1be073789b7d3",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by product names to filter products",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter options for products must from [best_selling,latest,top_rated,recommended]",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products with additional user-specific information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "description": "Current page of the paginated results",
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "first_page_url": {
                                            "description": "URL for the first page of results",
                                            "type": "string"
                                        },
                                        "last_page_url": {
                                            "description": "URL for the last page of results",
                                            "type": "string"
                                        },
                                        "next_page_url": {
                                            "description": "URL for the next page of results",
                                            "type": "string"
                                        },
                                        "prev_page_url": {
                                            "description": "URL for the previous page of results",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid search or filter parameters"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve a specific product by ID",
                "operationId": "cb06f46365fbb9c65892d84a25cfb29f",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "The unique identifier for the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "401": {
                        "description": "Unauthorized - Authentication is required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{product}/rate": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Rate a product",
                "operationId": "dafb2d482542749df0ab2255747e323e",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "The unique identifier for the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rate"
                                ],
                                "properties": {
                                    "rate": {
                                        "description": "Rating value between 1 and 5",
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product rating successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the operation",
                                            "type": "boolean"
                                        },
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation error for rate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "rate": {
                                                    "description": "Validation errors for the rate",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Authentication is required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "required": [
                    "id",
                    "name",
                    "price",
                    "quantity",
                    "category_id",
                    "store_id",
                    "rate",
                    "sales",
                    "created_at",
                    "updated_at",
                    "your_rate",
                    "your_favourite"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier for the product",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the product",
                        "type": "string"
                    },
                    "store_id": {
                        "description": "The ID of the store the product belongs to",
                        "type": "integer"
                    },
                    "price": {
                        "description": "The price of the product in the smallest currency unit (e.g., cents)",
                        "type": "integer"
                    },
                    "discount": {
                        "description": "Discount percentage for the product",
                        "type": "integer",
                        "default": 0,
                        "maximum": 100,
                        "minimum": 0
                    },
                    "quantity": {
                        "description": "The available quantity of the product",
                        "type": "integer"
                    },
                    "category_id": {
                        "description": "The ID of the category the product belongs to",
                        "type": "integer"
                    },
                    "expire_date": {
                        "description": "The expiration date of the product, if applicable",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "photo": {
                        "description": "The URL of the product's photo or image path",
                        "type": "string",
                        "nullable": true
                    },
                    "rate": {
                        "description": "The average rating of the product",
                        "type": "number",
                        "format": "float",
                        "default": 0,
                        "maximum": 5,
                        "minimum": 0
                    },
                    "sales": {
                        "description": "The total number of sales for the product",
                        "type": "integer",
                        "default": 0
                    },
                    "your_rate": {
                        "description": "The rating given by the authenticated user to this product, default is 0",
                        "type": "integer"
                    },
                    "your_favourite": {
                        "description": "Indicates whether the product is marked as a favourite by the authenticated user",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Timestamp of when the product was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp of when the product was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use 'Bearer <token>' in the Authorization header",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}